\name{fmd}
\docType{data} 
\alias{fmd}
\title{2001 food-and-mouth epidemic, north Cumbria (UK)}
\description{
This data set gives the spatial locations and reported times of food-and-mouth disease in north Cumbria (UK), 2001. It is of no scientific value, as it deliberately excludes confidential information on farms at risk in the study-region. It is included in the package purely as an illustrative example.
}
\usage{data(fmd)}
\format{A matrix containing (x,y,t) coordinates of the 648 observations.}
\author{ Edith Gabriel <edith.gabriel@univ-avignon.fr> }
\references{
Diggle, P., Rowlingson, B. and Su, T. (2005). Point process methodology for on-line spatio-temporal disease surveillance. Environmetrics, 16, 423--34.
}
\seealso{
 \code{\link{northcumbria}} for boundaries of the county of north Cumbria.
 }
\keyword{datasets}
