\name{kml_close}
\alias{kml_close}

\title{Closes a KML file connexion}

\description{Closes the KML file connexion initialized by \code{kml_open()}.}

\usage{kml_close()}

\details{This function detects and closes the KML file connexion as initialized by \code{kml_open{}}.}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\references{
%% ~put references to the literature/web site here ~
}

\author{Pierre Roudier and Dylan Beaudette}

\note{
%%  ~~further notes~~
}

\seealso{\code{\link{kml}}, \code{\link{kml_open}}, \code{\link{kml_layer}}}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
"kml_close"
}

\keyword{ ~KML }
