\name{northcumbria}
\docType{data} 
\alias{northcumbria}
\title{Polygon boundary of north Cumbria}
\description{
This data set gives the boundary of the county of north Cumbria (UK).
}
\usage{data(northcumbria)}
\format{A matrix containing (x,y) coordinates of the boundary.}
\author{ Edith Gabriel <edith.gabriel@univ-avignon.fr> }
\seealso{
 \code{\link{fmd}} for the space-time pattern of food-and-mouth disease in this county in 2001.
}

\keyword{datasets}