\name{plotKML-package}
\alias{plotKML-package}
\alias{plotKML}
\docType{package}
\title{Plotting of space-time objects in Google Earth}
\description{A suite of functions for converting sp-class objects into KML or KMZ documents for use in Google Earth.}
\details{
\tabular{ll}{
Package: \tab plotKML\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-05-25\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Tomislav Hengl (\email{tom.hengl@wur.nl}), Pierre Roudier (\email{pierre.roudier@landcare.nz}), Dylan Beaudette (\email{debeaudette@ucdavis.edu})\cr
Maintainer: Tomislav Hengl
}
\references{
\itemize{
\item Home URL (\url{http://plotkml.r-forge.r-project.org/})
\item See also: \strong{Google Earth Outreach} project (\url{http://earth.google.com/outreach/tutorials.html})
}
}
\note{This package has been developed as a part of the \href{http://isric.org/projects/global-soil-information-facilities-gsif}{\strong{Global Soil Information Facilities}} project, which is run jointly by the ISRIC Institute and collaborators.}
\keyword{ package }
\keyword{ GSIF }
