\name{SpatialSamplingPattern-class}
\docType{class}
\alias{SpatialSamplingPattern-class}
\alias{plotKML,SpatialSamplingPattern}
\title{A class for spatial samples produced using various \code{spsample} methods}
\description{A class containing input and output objects generated by some sampling optimisation algorithm. Objects of this type can be directly visualized in Google Earth by using the \code{\link{plotKML-method}}.}
\section{Slots}{
  \describe{    
    \item{\code{method}:}{object of class \code{"character"}; sampling optimisation method}
    \item{\code{pattern}:}{object of class \code{"SpatialPoints"}; sampling points}
    \item{\code{sp.domain}:}{object of class \code{"SpatialPolygonsDataFrame"}; spatial domain / strata}
  }
}
\section{Methods}{
  \describe{
	\item{plotKML}{\code{signature(obj = "SpatialSamplingPattern")}: plots generated sampling plan in Google Earth}
  }
}
\author{ Tomislav Hengl }
\seealso{ \code{\link{plotKML-method}}, \code{spcosa::spsample}, \code{\link{plotKML-method}}
}
\keyword{classes}
