\name{plotKML-package}
\alias{plotKML-package}
\docType{package}
\encoding{latin1}
\title{Visualization of spatial and spatio-temporal objects in Google Earth}
\description{A suite of functions for converting 2D and 3D spatio-temporal (sp, raster and spacetime package classes) objects into KML or KMZ documents for use in Google Earth.}
\details{
\tabular{ll}{
Package: \tab plotKML\cr
Type: \tab Package\cr
URL: \tab \url{http://plotkml.r-forge.r-project.org/}\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Tomislav Hengl (\email{tom.hengl@opengeohub.org}), Pierre Roudier (\email{pierre.roudier@landcare.nz}), Dylan Beaudette (\email{debeaudette@ucdavis.edu}), Edzer Pebesma (\email{edzer.pebesma@uni-muenster.de})
}
\references{
\itemize{
\item KML documentation (\url{http://code.google.com/apis/kml/documentation/})
\item \strong{Google Earth Outreach} project (\url{http://earth.google.com/outreach/tutorials.html})
\item Hengl, T., Roudier, P., Beaudette, D. and Pebesma, E. (2015) \href{http://www.jstatsoft.org/v63/i05/}{plotKML: Scientific Visualization of Spatio-Temporal Data}. Journal of Statistical Software, 63(5): 1--25.
}
}
\note{This package has been developed as a part of the Global Soil Information Facilities project, which is run jointly by the ISRIC Institute and collaborators. ISRIC is a non-profit organization with a mandate to serve the international community as custodian of global soil information and to increase awareness and understanding of the role of soils in major global issues.}
\keyword{spatial}